优化：自增ID字段升级为bigint-代码生成

+++ sz-common/sz-common-generator/src/main/java/com/sz/generator/core/util/GeneratorUtils.java
@@ -43,7 +43,7 @@ public class GeneratorUtils {
         return generatorTable;
     }
 
-    public static GeneratorTableColumn initColumnField(TableColumResult column, Integer tableId, int i) {
+    public static GeneratorTableColumn initColumnField(TableColumResult column, Long tableId, int i) {
         GeneratorTableColumn tableColumn = new GeneratorTableColumn();
         String dataType = getDbType(column.getColumnType());
         String columnName = column.getColumnName();
