+++ sz-common/sz-common-generator/src/main/java/com/sz/generator/service/impl/GeneratorTableColumnServiceImpl.java
@@ -29,14 +29,14 @@ public class GeneratorTableColumnServiceImpl extends ServiceImpl<GeneratorTableC
     }
 
     @Override
-    public List<GeneratorTableColumn> getTableColumnsByTableId(Integer tableId) {
+    public List<GeneratorTableColumn> getTableColumnsByTableId(Long tableId) {
         QueryWrapper wrapper = QueryWrapper.create().eq(GeneratorTableColumn::getTableId, tableId).orderBy(GeneratorTableColumn::getSort).asc();
         List<GeneratorTableColumn> list = list(wrapper);
         return list;
     }
 
     @Override
-    public List<GeneratorTableColumn> getTableColumnsByTableName(Integer tableId) {
+    public List<GeneratorTableColumn> getTableColumnsByTableName(Long tableId) {
         List<GeneratorTableColumn> list = QueryChain.of(mapper).eq(GeneratorTableColumn::getTableId, tableId).orderBy(GeneratorTableColumn::getSort).asc()
                 .list();
         return list;
